firmware: imx: remove prebuilt ATF binary
The prebuilt ATF binary is in fact a license compliance issue. While
the used BSD-3-clause license allows aggregation with the Barebox
binary, the chosen license requires us to ship the license text
itself for the source distribution and also for the binary form if
the built boot image contains the ATF binary.

As this is a restriction that only applies to some boot images,
depending on the Barebox configuration, it's non-trivial to do
proper license compliance from within the Barebox source.

Remove the prebuilt binary and rely on the external build system
or user to provide the correct binary. Most system build systems
already provide a proper way to handle the license dependency
issue.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>