fs: devfs: forbid truncation when cdev has no truncate operation
When a cdev doesn't have a truncate callback then forbid truncation
and fail with -EPERM.
Before this we had always failed with -ENOSPC in this situation.
We checked for f->fsdev->dev.num_resources being nonzero, but this
check was absolutely meaningless. It goes back to ancient times when
the resources of a device were automatically added to devfs.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>