scripts: imx-image: make read_file behaviour consistent
When read_file fails then it exits the program in most cases. Let's exit
the program when stat() fails aswell, so that we *always* exit the
program on failure. Drop all error handling for read_file since that's
no longer necessary and rename the function to xread_file to make
clear it won't fail.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>