Revert "images: Drop unnecessary fix_size"
This reverts commit 5c0e16591d9471a345b77a41fde76de34f301f6b.

The fix_size scripts is not necessary for newer ARM toolchains,
it is however necessary for the older ARM toolchains (gcc-5 and older).

The original reason to drop fix_size was that it doesn't work on MIPS.
With this patch we add the -i flag so that we ignore unknown images
and can succesfully build MIPS images.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>