ARM: rpi: switch emmc from sdhost to sdhci
BCM2837 has two mmc interfaces: sdhost and sdhci. On Raspberry Pi 3, sdhost is
normally used for SD card/MMC flash and sdhci for the wireless interface.
Barebox currently only has a driver for sdhci, so we disable the sdhost
interface and remap the pins so that the sdhci has access to the SD card/MMC
flash.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>