LICENSES: adopt Linux-like LICENSES directory structure
At the moment
  grep -r --exclude-dir=.git 'SPDX-License-Identifier:' | \
    grep -v 'GPL-2.0'

shows me 39 non-dually-licensed source code files (SoCFPGA) that have a
BSD-3-Clause license _identifier_.
There seems to be no barebox BSD-3-Clause license  _text_ however,
which runs afoul of the first clause of the license.

To account for this and future contributions which are licensed under
non-GPL-2.0-but-compatible licenses, create a new LICENSES directory
like the one Linux has, where licenses are located.

This also removes the jumptable exception inherited from U-Boot:

  > On 11/2/19 09:00, Sascha Hauer wrote:
  >> I don't think we need this [exception]. barebox doesn't have
  >> (and never had since the fork from U-Boot) these standalone
  >> applications. There is no jumptable in barebox. What we do have
  >> is modules support, but these are not covered by this text.

The LICENSES/exceptions directory is therefore empty as well as
the LICENSES/other directory, which might house new licenses in future.
They are added along this commit however, because upstream
checkpatch.pl introduced with a later commit depends on their
existence.

Fixes: a3ffa97f4 ("rename U-Boot-v2 project to barebox")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>