defaultenv: Convert init script to C
It's hard to get more complicated things right in hush. This commit
converts the /env/bin/init script to C code. With this we get a better
error handling and better control what is being done.

If /env/bin/init exists in the environment then it is still executed
instead of the corresponding C code.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>