arch: add KBUILD_DEFCONFIG to allow "make defconfig"
Currently, "make defconfig" is only supported for sandbox and x86.

I set KBUILD_DEFCONFIG to support "make defconfig" where I was able
to find a suitable one.

nios2 and openrisc have a single generic_defconfig, so it is definitely
good for KBUILD_DEFCONFIG.

For arm, mips, I assume qemu is a preferred choice over a real board.

blackfin has only one defconfig. Since it is inactive (it is gone in
Linux), so no more defconfig will come in.

I am not sure about ppc and riscv, so I did not touch them.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>