treewide: replace '---help---' with 'help' in Kconfig files
In Linux, '---help---' was deprecated in favor of 'help', and
this is checked by the recent checkpatch.pl

See Linux commit 84af7a6194e493fae312a2b7fa5a3b51f76d9282

The number of '---help---' is gradually decreasing in Linux, but there
are still lots. However, '---help---' will be completely killed when
the time comes.

Fortunately, there are only some in Barebox. Replacing them is not hard.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>