mvebuimg: new tool to create kwb version1 boot images
The purpose is the same as kwbimage, but the code is saner (in my eyes
at least). Currently it only supports creating v1 images, but can be
easily expanded to support v0 and extraction, too. The main advantage is
that the file name of the binary headers is passed via cmdline and so
the barebox build system can know about them without specifying them
twice.

Additionally it supports more than one binary header and doesn't create
broken images when running on a big endian platform.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>