ARM: rpi: use sdhost for eMMC access on CM3.
This commit removes the fix for the upstream devicetree where eMMC was remapped
to the sdhci interface. This is no longer needed since sdhost driver is now
functional in barebox. A similar change was done to the R.Pi 3 devicetree in
9036d1db.

In fact, this remap prevented CM3 from booting if barebox was compiled with
CONFIG_MCI_BCM283X_SDHOST (in rpi_defconfig since b093468c). In that case both
sdhost and sdhci drivers were active and tried to set their corresponsing alt
functions for the same set of GPIOs at the same time.

Signed-off-by: Tomaz Solc <tomaz.solc@tablix.org>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>