startup: Factor out the autoboot counter to separate function
The autoboot countdown is part of the init function. This patch factors
out this code to a separate function to allow boards to call it at a
different time. Also boards can set the autoboot state manually in case
they have its own idea how to interrupt the autoboot process.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>