pstore: Only capture log messages
With this pstore only captures barebox log messages printed with pr_*
and dev_*, but no longer anything printed with printf and friends.

When capturing the barebox output with pstore only the log messages are
of interest, but not the ones printed with printf and certainly not
the things typed interactively on the command line. These are logged
currently because we register pstore as a barebox console. Instead, hook
into pr_puts which only outputs the barebox log messages.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>