net: cpsw: adopt to upstream device tree changes
Upstream device trees no longer have a "cpsw-phy-sel" property to find
the phy_sel register, instead they have a child device of the pinctrl
node compatible to "ti,am3352-phy-gmii-sel". Also the "rmii-clock-ext"
property is no longer global to the cpsw but instead can be selected
per slave. To adopt to these changes take the short way out for now
and find the new node by its compatible and hardcode the "rmii-clock-ext"
setting (which is set to true in am33xx-l4.dtsi and not overwritten
by any board).

This makes the cpsw driver work again. Tested on Beaglebone black
board.

Fixes: 1dc748b3b2 ("dts: update to v5.1-rc1")
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>